[Refactor]: DTO 패키지 위치 이동
본문

[#?]
