+++ src/main/java/project/domain/auth/dto/request/SignUpRequest.java
@@ -1,4 +1,4 @@
-package project.domain.user.dto.login;
+package project.domain.auth.dto.request;
 
 import lombok.Getter;
 import lombok.Setter;
