+++ src/main/java/project/domain/auth/dto/response/SignUpResponse.java
@@ -1,4 +1,4 @@
-package project.domain.user.dto.login;
+package project.domain.auth.dto.response;
 
 import lombok.Getter;
